+++ tests/corpus/raw-baseline/arm64_O1/xxhash32.c
@@ -95,7 +95,7 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
                     }
                 }
                 {
-                    uint32_t tmp_lane_100000970_0_3f_1 = (uint32_t)X14_11;
+                    uint32_t tmp_lane_100000970_0_3f_1 = X14_11;
                     uint32_t tmp_2b380_13 = tmp_2a000_4 * (tmp_lane_100000970_0_3f_1 >> 15 ^ tmp_lane_100000970_0_3f_1);
                     uint32_t tmp_2b380_14 = tmp_2a000_13 * (tmp_2b380_13 >> 13 ^ tmp_2b380_13);
                     return tmp_2b380_14 >> 16 ^ tmp_2b380_14;
