+++ tests/corpus/raw-baseline/x64_O2/sdbm.c
@@ -6,7 +6,7 @@ uint64_t sym__sdbm(uint64_t RDI_0, uint64_t RSI_0)
     } else {
         uint64_t RDX_1;
         uint64_t RAX_1;
-        uint64_t RCX_2 = (uint64_t)(uint32_t)((uint32_t)RSI_0 & 3);
+        uint64_t RCX_2 = (uint64_t)((uint32_t)RSI_0 & 3);
         if (4 <= RSI_0) {
             RSI_0 &= (uint64_t)-0x4;
             RDX_1 = (uint64_t)0;
@@ -16,7 +16,7 @@ uint64_t sym__sdbm(uint64_t RDI_0, uint64_t RSI_0)
                 uint8_t tmp_11e00_3 = ((uint8_t*)RDI_0)[RDX_1 + 1];
                 uint8_t tmp_11e00_4 = ((uint8_t*)RDI_0)[RDX_1 + 2];
                 uint8_t tmp_11e00_5 = ((uint8_t*)RDI_0)[RDX_1 + 3];
-                RAX_1 = (uint64_t)(uint32_t)((uint32_t)tmp_11e00_5 + (uint32_t)((uint64_t)(int32_t)((uint32_t)tmp_11e00_4 + (uint32_t)((uint64_t)(int32_t)((uint32_t)tmp_11e00_3 + (uint32_t)((uint64_t)(int32_t)((uint32_t)tmp_11e00_2 + (uint32_t)((uint64_t)(int32_t)RAX_1 * 0x1003f)) * 0x1003f)) * 0x1003f)) * 0x1003f));
+                RAX_1 = (uint64_t)((uint32_t)tmp_11e00_5 + (uint32_t)((uint64_t)(int32_t)((uint32_t)tmp_11e00_4 + (uint32_t)((uint64_t)(int32_t)((uint32_t)tmp_11e00_3 + (uint32_t)((uint64_t)(int32_t)((uint32_t)tmp_11e00_2 + (uint32_t)((uint64_t)(int32_t)RAX_1 * 0x1003f)) * 0x1003f)) * 0x1003f)) * 0x1003f));
                 RDX_1 += 4;
                 if (RSI_0 == RDX_1) {
                     break;
@@ -33,7 +33,7 @@ uint64_t sym__sdbm(uint64_t RDI_0, uint64_t RSI_0)
                 RDX_6 = (uint64_t)0;
                 for (; ; ) {
                     uint8_t tmp_11e00_8 = ((uint8_t*)RDI_0)[RDX_6];
-                    RAX_1 = (uint64_t)(uint32_t)((uint32_t)tmp_11e00_8 + (uint32_t)((uint64_t)(int32_t)RAX_1 * 0x1003f));
+                    RAX_1 = (uint64_t)((uint32_t)tmp_11e00_8 + (uint32_t)((uint64_t)(int32_t)RAX_1 * 0x1003f));
                     RDX_6++;
                     if (RCX_2 == RDX_6) {
                         break;
